+++ projects/ontimize-web-ngx/src/lib/components/button/o-button.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit, ViewEncapsulation } from '@angular/core';
+import { ThemePalette } from '@angular/material';
+import { InputConverter } from '../../decorators/input-converter';
+
 
 export const DEFAULT_INPUTS_O_BUTTON = [
   'oattr: attr',
@@ -9,7 +12,11 @@ export const DEFAULT_INPUTS_O_BUTTON = [
   'icon',
   'svgIcon : svg-icon',
   'iconPosition: icon-position',
-  'image'
+  'image',
+  // enabled [yes|no|true|false]: Whether the button is enabled. Default: yes
+  'enabled',
+  // color: Theme color palette for the component.
+  'color'
 ];
 
 @Component({
@@ -27,12 +34,14 @@ export class OButtonComponent implements OnInit {
   protected static DEFAULT_TYPE = 'STROKED';
 
   protected oattr: string;
-  olabel: string;
+  public olabel: string;
   protected otype: string;
-  icon: string;
-  svgIcon: string;
-  iconPosition: string; // left (default), top, TODO: right, bottom?
-  image: string;
+  public icon: string;
+  public svgIcon: string;
+  public iconPosition: string; // left (default), top, TODO: right, bottom?
+  public image: string;
+  @InputConverter() enabled: boolean = true;
+  public color: ThemePalette;
 
   constructor() {
     this.otype = OButtonComponent.DEFAULT_TYPE;