+++ projects/ontimize-web-ngx/src/lib/components/input/date-range/o-daterange-input.component.ts
@@ -208,7 +208,7 @@ export class ODateRangeInputComponent extends OFormDataComponent implements OnDe
 
   updateElement() {
     let chosenLabel: any;
-    if (Util.isDefined(this.value.value) && !this.isObjectDataRangeNull(this.value)) {
+    if (Util.isDefined(this.value) && Util.isDefined(this.value.value) && !this.isObjectDataRangeNull(this.value)) {
       if (this.value.value[this.pickerDirective.startKey] && this.value.value[this.pickerDirective.endKey]) {
         this.value.value[this.pickerDirective.startKey] = this.ensureDateRangeValue(this.value.value[this.pickerDirective.startKey], this._valueType);
         this.value.value[this.pickerDirective.endKey]  = this.ensureDateRangeValue(this.value.value[this.pickerDirective.endKey], this._valueType);
@@ -219,6 +219,7 @@ export class ODateRangeInputComponent extends OFormDataComponent implements OnDe
       }
     } else {
       chosenLabel = null;
+      this.pickerDirective.value = undefined;
     }
     this.pickerDirective._el.nativeElement.value = chosenLabel;
   }