+++ projects/ontimize-web-ngx/src/lib/components/table/extensions/header/o-table-header-components.ts
@@ -7,6 +7,7 @@ import { OTableMenuComponent } from './table-menu/o-table-menu.component';
 import { OTableOptionComponent } from './table-option/o-table-option.component';
 import { OTableQuickfilterComponent } from './table-quickfilter/o-table-quickfilter.component';
 import { OTableColumnsFilterColumnComponent } from './table-columns-filter/columns/o-table-columns-filter-column.component';
+import { OTableHeaderColumnFilterIconComponent } from './table-header-column-filter-icon/o-table-header-column-filter-icon.component';
 
 export const O_TABLE_HEADER_COMPONENTS = [
   OTableButtonComponent,
@@ -17,5 +18,6 @@ export const O_TABLE_HEADER_COMPONENTS = [
   OTableOptionComponent,
   OTableInsertableRowComponent,
   OTableQuickfilterComponent,
-  OTableColumnsFilterColumnComponent
+  OTableColumnsFilterColumnComponent,
+  OTableHeaderColumnFilterIconComponent
 ];