+++ projects/ontimize-web-ngx/src/lib/components/table/o-table.module.ts
@@ -25,6 +25,8 @@ import { OTableRowClassPipe } from './extensions/pipes/o-table-row-class.pipe';
 import { OTableRowDirective } from './extensions/row/o-table-row.directive';
 import { OMatSortModule } from './extensions/sort/o-mat-sort-module';
 import { OTableComponent } from './o-table.component';
+import { OTableRowExpandableComponent } from './extensions/row/table-row-expandable/o-table-row-expandable.component';
+import { PortalModule } from '@angular/cdk/portal';
 
 @NgModule({
   declarations: [
@@ -40,13 +42,15 @@ import { OTableComponent } from './o-table.component';
     ...O_TABLE_CELL_EDITORS,
     ...O_TABLE_DIALOGS,
     ...O_TABLE_HEADER_COMPONENTS,
-    ...O_TABLE_FOOTER_COMPONENTS
+    ...O_TABLE_FOOTER_COMPONENTS,
+    OTableRowExpandableComponent
   ],
   imports: [
     CommonModule,
     OSharedModule,
     CdkTableModule,
     DragDropModule,
+    PortalModule,
     OContextMenuModule,
     ObserversModule,
     OMatSortModule,
@@ -65,7 +69,8 @@ import { OTableComponent } from './o-table.component';
     ...O_TABLE_HEADER_COMPONENTS,
     ...O_TABLE_CELL_RENDERERS,
     ...O_TABLE_CELL_EDITORS,
-    ...O_TABLE_FOOTER_COMPONENTS
+    ...O_TABLE_FOOTER_COMPONENTS,
+    OTableRowExpandableComponent
   ],
   entryComponents: [
     OTableColumnAggregateComponent,